# contracts/openapi/openapi-v1.0.yaml
...
openapi: 3.1.0
info:
  title: 学习打卡系统 API
  version: 1.0.0
  description: >
    学习打卡系统 V1.0 契约。状态枚举、转换与时间规则见
    `memory-bank/state-machine-and-time-rules-V1.0.md`。需求与决策追踪见
    `memory-bank/requirements-tracking-V1.0.md`。
servers:
  - url: /api/v1
...
paths:
...
  /tasks/{taskId}/items:
    get:
      tags: [items]
      summary: 条目列表
      operationId: listItems
      description: 全量条目分页，按 sortOrder 升序稳定排序；含已归档（可筛）。
      security:
        - sessionCookie: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 条目分页。
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, requestId]
                properties:
                  data:
                    $ref: '#/components/schemas/ItemPage'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [items]
      summary: 手工新增条目
      operationId: createItem
      description: 自动分配稳定序号；校验标题必填、链接格式；校验任务为清单型、归属、状态。
      security:
        - sessionCookie: []
        - csrf: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreateRequest'
      responses:
        '201':
          $ref: '#/components/responses/Success'
        '422':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks/{taskId}/items/paste-preview:
    post:
      tags: [items]
      summary: 批量粘贴预览
      operationId: pastePreview
      description: 每行一条目；预览返回总数/有效/错误行。确认前不写正式条目。
      security:
        - sessionCookie: []
        - csrf: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PastePreviewRequest'
      responses:
        '200':
          description: 预览结果。
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, requestId]
                properties:
                  data:
                    $ref: '#/components/schemas/PastePreviewResponse'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tasks/{taskId}/items/paste-confirm:
    post:
      tags: [items]
      summary: 确认批量粘贴
      operationId: pasteConfirm
      description: 追加到既有条目之后，序号唯一。幂等键防重复确认。
      security:
        - sessionCookie: []
        - csrf: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasteConfirmRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /items/{itemId}:
    get:
      tags: [items]
      summary: 条目详情
      operationId: getItem
      security:
        - sessionCookie: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 条目详情。
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, requestId]
                properties:
                  data:
                    $ref: '#/components/schemas/ItemView'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [items]
      summary: 编辑条目
      operationId: updateItem
      description: 有历史/已完成条目禁止硬删除，编辑需版本号。
      security:
        - sessionCookie: []
        - csrf: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
  # ============ 今日计划 / 分配（S3） ============
  /tasks/{taskId}/today-items:
    get:
      tags: [items]
      summary: 今日计划条目
      operationId: getTodayItems
      description: 惰性创建当日打卡与分配；目标按剩余量缩减；未完成顺延优先。返回归属是否计入今日。
      security:
        - sessionCookie: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 今日计划。
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, requestId]
                properties:
                  data:
                    $ref: '#/components/schemas/TodayItemsResponse'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============ 题解（S4） ============
...
  /tasks/{taskId}/imports/xlsx/preview:
    post:
      tags: [imports]
      summary: 预览 xlsx 导入
      operationId: previewXlsxImport
      description: 以 multipart/form-data 上传 xlsx；服务端在请求内同步用 POI 解析固定列 title、content、analysis、link、order，返回候选项和错误行。预览不写正式条目，不创建导入批次或后台任务。
      security:
        - sessionCookie: []
        - csrf: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/XlsxPreviewRequest'
      responses:
        '200':
          description: xlsx 解析预览。
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, requestId]
                properties:
                  data:
                    $ref: '#/components/schemas/XlsxPreviewResponse'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/BadRequest'
  /tasks/{taskId}/imports/xlsx/confirm:
    post:
      tags: [imports]
      summary: 确认 xlsx 导入
      operationId: confirmXlsxImport
      description: 在单个事务内按预览候选执行标题去重并追加正式条目；默认跳过重复标题，不覆盖已有条目。无异步批次、轮询、拆分、合并或覆盖操作。重复提交以 Idempotency-Key 返回首次结果。
      security:
        - sessionCookie: []
        - csrf: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/XlsxConfirmRequest'
      responses:
        '200':
          description: xlsx 确认入库结果。
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, requestId]
                properties:
                  data:
                    $ref: '#/components/schemas/XlsxConfirmResponse'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BadRequest'


  # ============ 导出（S9） ============
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION_ID
      description: HttpOnly、Secure、SameSite=Lax 会话 Cookie。由后端签发。
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 所有写请求必须携带，token 从 GET /auth/csrf 获取，仅在内存。
  schemas:
...
    EnvelopeSuccess:
      type: object
      additionalProperties: false
      required: [data, requestId]
      properties:
        data: {}
        requestId:
          type: string
          description: 请求 ID，贯穿日志，用于故障定位。
    EnvelopeError:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: 稳定错误码，前端据此分类，不解析 message 文本。
            message:
              type: string
              description: 用户可理解的错误信息。
        requestId:
          type: string

    # ============ 认证 ============
...
    TaskType:
      type: string
      enum: [checklist]
    TaskStatus:
      type: string
      enum: [draft, active]
      description: 删除为物理删除；结束状态由 endDate 在任务时区下派生，不新增持久值。
    ScheduleType:
      type: string
      enum: [daily]
...
    TaskView:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - type
        - status
        - scheduleType
        - startDate
        - timezone
        - dailyTargetCount
        - ended
        - itemCount
        - completedItemCount
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/TaskType'
        status:
          $ref: '#/components/schemas/TaskStatus'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        scheduleType:
          $ref: '#/components/schemas/ScheduleType'
        timezone:
          type: string
        dailyTargetCount:
          type: integer
          minimum: 1
        ended:
          type: boolean
          description: 派生字段：active 且 endDate 早于任务时区今日时为 true，不是持久状态。
        itemCount:
          type: integer
          minimum: 0
        completedItemCount:
          type: integer
          minimum: 0
...
    ItemStatus:
      type: string
      enum: [pending, completed]
    ItemView:
      type: object
      additionalProperties: false
      required: [id, taskId, title, sortOrder, status]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
          nullable: true
          description: 题目正文。
        analysis:
          type: string
          nullable: true
          description: 题库解析，非用户题解。有值才显示"查看解析"。
        externalUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        status:
          $ref: '#/components/schemas/ItemStatus'
        sourcePageStart:
          type: integer
          nullable: true
        sourcePageEnd:
          type: integer
          nullable: true
        parseConfidence:
          type: number
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
    ItemPage:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemView'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    ItemCreateRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        content:
          type: string
        externalUrl:
          type: string
    ItemUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        externalUrl:
          type: string
    PastePreviewRequest:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text:
          type: string
          description: 每行一个条目的文本。
    PastePreviewResponse:
      type: object
      additionalProperties: false
      required: [totalLines, validLines, errorLines, candidates]
      properties:
        totalLines:
          type: integer
        validLines:
          type: integer
        errorLines:
          type: array
          items:
            $ref: '#/components/schemas/PasteErrorLine'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/PasteCandidate'
    PasteErrorLine:
      type: object
      additionalProperties: false
      required: [lineNumber, reason]
      properties:
        lineNumber:
          type: integer
        reason:
          type: string
    PasteCandidate:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
          nullable: true
    PasteConfirmRequest:
      type: object
      additionalProperties: false
      required: [candidateTitles]
      properties:
        candidateTitles:
          type: array
          items:
            type: string
          description: 确认写入的条目标题列表，按顺序追加。

    # ============ 今日计划 / 分配 ============
    TodayItem:
      type: object
      additionalProperties: false
      required: [item, assigned, belongsToToday]
      properties:
        item:
          $ref: '#/components/schemas/ItemView'
        assigned:
          type: boolean
          description: 是否今日分配项。
        belongsToToday:
          type: boolean
          description: 完成归属是否计入今日计划（非今日=只计总进度）。DEC-11。
    TodayItemsResponse:
      type: object
      additionalProperties: false
      required: [task, plannedCount, completedCount, items]
      properties:
        task:
          $ref: '#/components/schemas/TaskView'
        plannedCount:
          type: integer
          description: 今日目标（已按剩余量缩减）。
        completedCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodayItem'

    # ============ 题解 ============
...
    XlsxPreviewRequest:
      type: object
      additionalProperties: false
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: 仅支持 .xlsx；服务端校验 MIME、扩展名和内容并同步解析。
    XlsxPreviewResponse:
      type: object
      additionalProperties: false
      required: [totalRows, validRows, errorRows, candidates]
      properties:
        totalRows:
          type: integer
          minimum: 0
        validRows:
          type: integer
          minimum: 0
        errorRows:
          type: array
          items:
            $ref: '#/components/schemas/XlsxErrorRow'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/XlsxCandidate'
    XlsxErrorRow:
      type: object
      additionalProperties: false
      required: [rowNumber, reason]
      properties:
        rowNumber:
          type: integer
          minimum: 1
        reason:
          type: string
    XlsxCandidate:
      type: object
      additionalProperties: false
      required: [title, content, analysis, link, order, duplicate]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        content:
          type: string
          nullable: true
        analysis:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        order:
          type: integer
          nullable: true
        duplicate:
          type: boolean
          description: 标题规范化后与当前任务已有条目重复；确认默认跳过。
    XlsxConfirmRequest:
      type: object
      additionalProperties: false
      required: [candidates]
      properties:
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/XlsxConfirmCandidate'
    XlsxConfirmCandidate:
      type: object
      additionalProperties: false
      required: [title, action]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        content:
          type: string
          nullable: true
        analysis:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        order:
          type: integer
          nullable: true
        action:
          type: string
          enum: [skip, keep_new]
          description: 重复标题默认 skip；keep_new 仅追加新条目，不覆盖已有条目。
    XlsxConfirmResponse:
      type: object
      additionalProperties: false
      required: [createdCount, skippedCount]
      properties:
        createdCount:
          type: integer
          minimum: 0
        skippedCount:
          type: integer
          minimum: 0

    # ============ 导出 ============
...
  responses:
    Success:
      description: 成功，业务数据在 data 中。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeSuccess'
    BadRequest:
      description: 422 字段校验失败或 400。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeError'
    Conflict:
      description: 409 业务冲突，例如同一用户下任务名称重复或非法状态转换。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeError'
...
    Unauthorized:
      description: 401 会话失效。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeError'
...